+++ mitra_models/src/properties/constants.rs
@@ -1,5 +1,8 @@
 pub const INSTANCE_ED25519_SECRET_KEY: &str = "instance_ed25519_secret_key";
 pub const INSTANCE_RSA_SECRET_KEY: &str = "instance_rsa_secret_key";
-pub const FEDERATED_TIMELINE_RESTRICTED: &str = "federated_timeline_restricted";
+
+#[deprecated]
 pub const ETHEREUM_CURRENT_BLOCK: &str = "ethereum_current_block";
+
+pub const FEDERATED_TIMELINE_RESTRICTED: &str = "federated_timeline_restricted";
 pub const FILTER_KEYWORDS: &str = "filter_keywords";