+++ mitra_models/src/properties/queries.rs
@@ -29,7 +29,7 @@ pub async fn set_internal_property(
     Ok(())
 }
 
-pub async fn get_internal_property_json(
+async fn get_internal_property_json(
     db_client: &impl DatabaseClient,
     name: &str,
 ) -> Result<Option<JsonValue>, DatabaseError> {
@@ -64,8 +64,26 @@ pub async fn get_internal_property<T: DeserializeOwned>(
     Ok(maybe_value)
 }
 
+pub async fn get_internal_properties_json(
+    db_client: &impl DatabaseClient,
+) -> Result<JsonValue, DatabaseError> {
+    let row = db_client.query_one(
+        "
+        SELECT COALESCE(
+            jsonb_object_agg(property_name, property_value),
+            '{}'
+        ) AS property_map
+        FROM internal_property
+        ",
+        &[],
+    ).await?;
+    let properties = row.try_get("property_map")?;
+    Ok(properties)
+}
+
 #[cfg(test)]
 mod tests {
+    use serde_json::json;
     use serial_test::serial;
     use crate::database::test_utils::create_test_database;
     use super::*;
@@ -80,5 +98,17 @@ mod tests {
         let db_value: u32 = get_internal_property(db_client, name).await
             .unwrap().unwrap_or_default();
         assert_eq!(db_value, value);
+    }
+
+    #[tokio::test]
+    #[serial]
+    async fn test_get_internal_properties_json() {
+        let db_client = &create_test_database().await;
+        let properties = get_internal_properties_json(db_client).await.unwrap();
+        assert_eq!(properties, json!({}));
+
+        set_internal_property(db_client, "myproperty", &100).await.unwrap();
+        let properties = get_internal_properties_json(db_client).await.unwrap();
+        assert_eq!(properties, json!({"myproperty": 100}));
     }
 }